requirements.txt
Django==5.2.12
asgiref==3.8.1
sqlparse==0.5.5
tzdata==2025.3

djangorestframework==3.15.2
django-filter==24.2
django-cors-headers==4.4.0

django-environ==0.11.2
django-redis==5.4.0
redis==5.0.1

gunicorn

mysqlclient==2.2.4

django-storages==1.14.4
boto3==1.34.162


PyJWT==2.8.0

pillow==10.4.0

python-dateutil==2.9.0.post0
typing_extensions==4.12.2
six==1.17.0
urllib3==2.2.2
